Move tests into respective folders
