--- dbcollection/tests/test_funs.py
@@ -1,29 +1,0 @@
-"""
-Test dbcollection/dataset/funs.py.
-"""
-
-
-import os
-import sys
-import pytest
-from dbcollection.datasets import funs
-
-
-def test_fetch_dataset_constructor__succeed():
-    pass
-
-def test_fetch_dataset_constructor__fail():
-    pass
-
-def test_setup_dataset_constructor():
-    pass
-
-def test_exists():
-    pass
-
-def test_download():
-    pass
-
-def test_process():
-    pass
-
